Get a list of suggestions for geolocations in NL

nl_suggest returns a list of suggestions for a location description.

nl_suggest(q, ..., rows = NULL, type = NULL, verbose = FALSE,
  fl = NULL, sort = NULL, qf = NULL, bq = NULL)

Arguments

q
search terms that should be geolocated

...
parameters passed to geolocation service

rows
maximum number of rows to be returned. Default is 10.

type
restrict type of geolocation to a type (see details for possible types)

verbose
Should the functions print messages on what it is retrieving.

fl
the columns that should be returned (aka select on columms of result)

sort
how the data should be sorted

qf
the fields that should be queried

bq
the boosting of the query.

Value

The result of the pdok suggest webservice converted to a R list object.

Details

type can be one or more of the following: "provincie", "gemeente" , "woonplaats", "weg", "postcode", "adres", "perceel", "hectometerpaal", "wijk", "buurt", "waterschapsgrens", "appartementsrecht".

Examples

sug <- nl_suggest("Henri Faasdreef")

# how many objects have a score?
sug$response$numFound

#> [1] 12

# get suggestions
sug$response$docs

#>     type                               weergavenaam
#> 1    weg             Henri Faasdreef, 's-Gravenhage
#> 2  adres        Henri Faasdreef 194T, 's-Gravenhage
#> 3  adres  Henri Faasdreef 194, 2492JP 's-Gravenhage
#> 4  adres  Henri Faasdreef 200, 2492JP 's-Gravenhage
#> 5  adres  Henri Faasdreef 218, 2492JP 's-Gravenhage
#> 6  adres  Henri Faasdreef 222, 2492JP 's-Gravenhage
#> 7  adres Henri Faasdreef 222A, 2492JP 's-Gravenhage
#> 8  adres  Henri Faasdreef 226, 2492JP 's-Gravenhage
#> 9  adres  Henri Faasdreef 228, 2492JP 's-Gravenhage
#> 10 adres  Henri Faasdreef 300, 2492JP 's-Gravenhage
#>                                      id    score
#> 1  weg-f633e85f07eda4e68a00fb13f9d128f5 32.06973
#> 2  adr-c3d9770bcc6a2dc762475d5fe5f619cf 22.83119
#> 3  adr-e1e06434f4fdc53db3bf29ea61304b7a 19.93772
#> 4  adr-6c3a7c000571ee07da9f48e7ce7cc9ae 19.93772
#> 5  adr-d0613907ca3ff239b0f42f06f1ecd38c 19.93772
#> 6  adr-188be71cc18d276712a65d814de8aa05 19.93772
#> 7  adr-f91cc969743132a5bef901e9652aa665 19.93772
#> 8  adr-d4d551eeea46611e731ba6fe12008bef 19.93772
#> 9  adr-e17eb123835906df64da85a1b2139ecb 19.93772
#> 10 adr-2abdec9b964c6df00be9d0af129f9fed 19.93772